Law-Lens-main/requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf
pdf2image
pytesseract
transformers
nest_asyncio
PyPDF2
pypdfium2
torch
Pillow
pandas
python-dotenv
spacy
tqdm
keybert
sentencepiece
groq
sentence-transformers
scikit-learn